requirements.txt
Django
Django-suit
django-wysiwyg-redactor
Pillow
psycopg2